refactor(channels): extract useMetaSignup from the connect modal
MetaChannelConnect.vue was 559 lines carrying three different logins on top of
the manual credentials form and agent assignment. The logins have nothing to do
with the form — they only need to hand back a connected account — so they move
to a composable and the component goes back to presentation.

Dropped isSignupChannel/SIGNUP_CHANNELS on the way: every channel this modal
handles now offers one-click connect, so the guard could no longer be false.

MetaChannelConnect 559 -> 370 lines, composable 256. No behaviour change:
typecheck clean and all 194 frontend tests pass untouched.
